payouts/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
requests==2.32.3
python-dotenv==1.1.0
apscheduler==3.11.0
sqlalchemy==2.0.40
alembic==1.15.2
psycopg[binary]==3.2.6
pydantic==2.11.3
pytest==8.3.5
httpx==0.28.1
